+++ src/components/router.js
@@ -1,7 +1,8 @@
 import React from "react";
 import { lazy } from "react";
 import { Routes, Route, useLocation } from "react-router-dom";
-import SignInComponent from './signup';
+import SignInComponent from '.././pages/signup';
+//import UpdateComponent from '.././pages/update';
 
 
 export function RouterPaths() {
@@ -10,6 +11,7 @@ export function RouterPaths() {
     
       <Routes location={location} key={location.pathname}>
         <Route exact path="/signup" element={<SignInComponent></SignInComponent>} />
+        
         
       </Routes>
